Practica gestión de archivos

+++ demo-ep/src/main/java/com/example/eprunonosa/demo_ep/services/CategoryService.java
@@ -2,19 +2,14 @@ package com.example.eprunonosa.demo_ep.services;
 
 import com.example.eprunonosa.demo_ep.model.Category;
 import com.example.eprunonosa.demo_ep.repository.CategoryRepository;
-import lombok.AllArgsConstructor;
-import lombok.Data;
-import lombok.NoArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
 @Service
-@AllArgsConstructor
-@NoArgsConstructor
-@Data
 public class CategoryService {
+
     @Autowired
     private CategoryRepository categoryRepository;
 
@@ -27,10 +22,14 @@ public class CategoryService {
     }
 
     public Category findById(Long id) {
-        return categoryRepository.findById(id).get();
+        return categoryRepository.findById(id).orElse(null);
     }
 
-    public void deleteById(Long id) {
-        categoryRepository.deleteById(id);
+    public boolean deleteById(Long id) {
+        if (categoryRepository.existsById(id)) {
+            categoryRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
 }
